requirements.txt
streamlit
google-generativeai
pandas
PyPDF2